+++ lib/view_component.rb
@@ -16,7 +16,9 @@ module ViewComponent
   autoload :Preview
   autoload :PreviewTemplateError
   autoload :TestHelpers
+  autoload :SystemTestHelpers
   autoload :TestCase
+  autoload :SystemTestCase
   autoload :TemplateError
   autoload :Translatable
 end