+++ docs/CHANGELOG.md
@@ -10,6 +10,10 @@ nav_order: 5
 
 ## main
 
+* Add ability to pass in explicit preview_path to preview generator.
+
+    *Erinna Chen*
+
 ## 2.78.0
 
 * Support variants with dots in their names.