+++ docs/CHANGELOG.md
@@ -10,6 +10,10 @@ nav_order: 5
 
 ## main
 
+* Add `with_rendered_component_path` helper for writing component system tests.
+
+    *Edwin Mak*
+
 * Include gem name and deprecation horizon in every deprecation message.
 
     *Jan Klimo*